+++ db/updates/update_8.8-8.9.sql
@@ -1,0 +1,1 @@
+UPDATE `settings` SET `pv_online_api_url` = 'https://online.perfumersvault.com/api-data/api.php' WHERE `settings`.`id` = 1;