+++ releasenotes.md
@@ -1,4 +1,5 @@
 Whats New in v8.9
 --------------------------
+- Add option to wipe ingredients database
 - Various bug fixes and improvements
 - For full details please refer to the CHANGELOG