+++ releasenotes.md
@@ -1,5 +1,6 @@
 Whats New in v8.9
 --------------------------
-- Add option to wipe ingredients database
+- Add option to wipe out all ingredients
+- Add option to wipe out all formulas
 - Various bug fixes and improvements
 - For full details please refer to the CHANGELOG