+++ releasenotes.md
@@ -1,5 +1,6 @@
 Whats New in v8.9
 --------------------------
+- Fix a bug failing to populate soem fields when adding a supplier
 - Update to PV Online v2
 - Add option to wipe out all ingredients
 - Add option to wipe out all formulas