+++ releasenotes.md
@@ -1,4 +1,6 @@
 Whats New in v9.0
 --------------------------
+- Search in formula for ingredients for multiples using comma separated values
+
 - Various bug fixes and improvements
 - For full details please refer to the CHANGELOG