+++ releasenotes.md
@@ -1,6 +1,4 @@
-Whats New in v9.0
+Whats New in v9.1
 --------------------------
-- Search in formula for ingredients for multiples using comma separated values
-- Batch import from PV Online has been deprecated and removed, you can still import single materials
 - Various bug fixes and improvements
 - For full details please refer to the CHANGELOG